added logic for creating event without integrations

+++ server/routes/private/events/controller.ts
@@ -1,6 +1,7 @@
 import { NextFunction, Request, Response } from "express";
 import { Readable } from "stream";
 import moment from "moment";
+import { v4 as uuidv4 } from "uuid";
 import { DATE_FORMAT, Services } from "../../../constants/services";
 import { Integration, User, Event } from "../../../db/sequelize";
 import connectionSchema from "../../../integrations/ConnectionSchema";
@@ -15,113 +16,124 @@ export const SynchronizeCalendar = async (
 ) => {
   const { id } = req.user as User;
 
-  const integrations: Integration[] = await Integration.findAll({
-    where: {
-      user_id: id,
-    },
-  });
+  try {
+    const integrations: Integration[] = await Integration.findAll({
+      where: {
+        user_id: id,
+      },
+    });
 
-  await Event.destroy({
-    where: {
-      user_id: id,
-    },
-  });
+    await Event.destroy({
+      where: {
+        user_id: id,
+      },
+    });
 
-  async function* getAllEvents() {
-    for (let i = 0; i < integrations.length; i++) {
-      const events: any[] = [];
-      const item = integrations[i];
-      const key = getKeyByValue(item.type as any) as keyof typeof Services;
+    async function* getAllEvents() {
+      for (let i = 0; i < integrations.length; i++) {
+        const events: any[] = [];
+        const item = integrations[i];
+        const key = getKeyByValue(item.type as any) as keyof typeof Services;
 
-      const integration = await Integration.findOne({
-        where: {
-          user_id: id,
-          type: item.type,
-        },
-      });
-
-      let data: any = [];
-
-      try {
-        data = await getArrayFromStream(
-          Readable.from(
-            connectionSchema[key].getEvents(
-              integration?.access_token as string,
-              moment().format(DATE_FORMAT),
-              moment().add(7, "d").format(DATE_FORMAT)
-            )
-          )
-        );
-      } catch (e) {
-        let token;
-        try {
-          token = await connectionSchema[key].exchangeToken(item.refresh_token);
-          console.log(
-            "🚀 ~ file: controller.ts ~ line 108 ~ integrations.map ~ token",
-            token
-          );
-        } catch (e: any) {
-          console.log("e: ", e);
-          return next(ApiError.badRequest("Internal error"));
-        }
-        data = await getArrayFromStream(
-          Readable.from(
-            connectionSchema[key].getEvents(
-              token?.data.access_token,
-              moment().format(DATE_FORMAT),
-              moment().add(7, "d").format(DATE_FORMAT)
-            )
-          )
-        );
-
-        if (key === "microsoftCalendar") {
-          await Integration.update(
-            {
-              access_token: token?.data.access_token,
-              refresh_token: token?.data.refresh_token,
-            },
-            {
-              where: {
-                user_id: id,
-              },
-            }
-          );
-        }
-      }
-
-      data.map((item: any) => {
-        events.push(...item);
-      });
-
-      for (let i = 0; i < events.length; i++) {
-        await Event.findOrCreate({
+        const integration = await Integration.findOne({
           where: {
-            event_id: events[i].id,
-            start_time: moment.utc(events[i].start.dateTime).valueOf(),
-            end_time: moment.utc(events[i].end.dateTime).valueOf(),
-          },
-          defaults: {
-            subject: events[i].subject,
-            start_time: moment.utc(events[i].start.dateTime).valueOf(),
-            end_time: moment.utc(events[i].end.dateTime).valueOf(),
             user_id: id,
-            integration_id: integration?.id,
-            body: "",
-            event_id: events[i].id,
+            type: item.type,
           },
-        } as any);
+        });
+
+        let data: any = [];
+
+        try {
+          data = await getArrayFromStream(
+            Readable.from(
+              connectionSchema[key].getEvents(
+                integration?.access_token as string,
+                moment().format(DATE_FORMAT),
+                moment().add(7, "d").format(DATE_FORMAT)
+              )
+            )
+          );
+        } catch (e) {
+          let token;
+          try {
+            token = await connectionSchema[key].exchangeToken(
+              item.refresh_token
+            );
+            console.log(
+              "🚀 ~ file: controller.ts ~ line 108 ~ integrations.map ~ token",
+              token
+            );
+          } catch (e: any) {
+            console.log("e: ", e);
+            return next(ApiError.badRequest("Internal error"));
+          }
+
+          try {
+            data = await getArrayFromStream(
+              Readable.from(
+                connectionSchema[key].getEvents(
+                  token?.data.access_token,
+                  moment().format(DATE_FORMAT),
+                  moment().add(7, "d").format(DATE_FORMAT)
+                )
+              )
+            );
+          } catch (e) {
+            return next(ApiError.badRequest("Internal error"));
+          }
+
+          if (key === "microsoftCalendar") {
+            await Integration.update(
+              {
+                access_token: token?.data.access_token,
+                refresh_token: token?.data.refresh_token,
+              },
+              {
+                where: {
+                  user_id: id,
+                },
+              }
+            );
+          }
+        }
+
+        data.map((item: any) => {
+          events.push(...item);
+        });
+
+        for (let i = 0; i < events.length; i++) {
+          await Event.findOrCreate({
+            where: {
+              event_id: events[i].id,
+              start_time: moment.utc(events[i].start.dateTime).valueOf(),
+              end_time: moment.utc(events[i].end.dateTime).valueOf(),
+            },
+            defaults: {
+              subject: events[i].subject,
+              start_time: moment.utc(events[i].start.dateTime).valueOf(),
+              end_time: moment.utc(events[i].end.dateTime).valueOf(),
+              user_id: id,
+              integration_id: integration?.id,
+              body: "",
+              event_id: events[i].id,
+            },
+          } as any);
+        }
+
+        yield events;
       }
-
-      yield events;
     }
-  }
 
-  let allEvents: any = [];
-  for await (let events of getAllEvents()) {
-    allEvents.push(...events);
-  }
+    let allEvents: any = [];
+    for await (let events of getAllEvents()) {
+      allEvents.push(...events);
+    }
 
-  return res.json(allEvents);
+    return res.json(allEvents);
+  } catch (e) {
+    return next(ApiError.internal("Internal error"));
+  }
 };
 
 export const getAllEvents = async (
@@ -150,60 +162,37 @@ export const CreateEvents = async (
   next: NextFunction
 ) => {
   const { types, startTime, endTime, subject, description } = req.body;
-  console.log("🚀 ~ file: controller.ts ~ line 153 ~ types", types);
   const { id: userId } = req.user as User;
 
-  const integrations: Integration[] = await Integration.findAll({
-    where: {
-      user_id: userId,
-    },
-  });
-
-  for (let i = 0; i < types.length; i++) {
-    const item: keyof typeof Services = types[i];
-    const { access_token, refresh_token, id } = integrations.find(
-      (integration) => integration.type === Services[item]
-    ) as Integration;
-
-    try {
-      const event: any = await connectionSchema[item].createEvent(
-        access_token,
-        {
-          description,
-          endTime,
-          startTime,
-          subject,
-        }
-      );
-      console.log("🚀 ~ file: controller.ts ~ line 176 ~ event", event.data);
-
-      const normalizedEvent = connectionSchema[item].normalize(event.data);
-
-      const {
-        end,
-        id: eventId,
-        start,
-        subject: normalizedSubject,
-      } = normalizedEvent;
-
-      const newEvent = new Event({
-        subject: normalizedSubject,
-        start_time: moment(start.dateTime).valueOf(),
-        end_time: moment(end.dateTime).valueOf(),
+  try {
+    const integrations: Integration[] = await Integration.findAll({
+      where: {
         user_id: userId,
-        integration_id: id,
+      },
+    });
+
+    if (types.length == 0) {
+      const newEvent = new Event({
+        subject: subject,
+        start_time: moment(startTime).valueOf(),
+        end_time: moment(endTime).valueOf(),
+        user_id: userId,
         body: "",
-        event_id: eventId,
+        event_id: uuidv4(),
       } as any);
 
       await newEvent.save();
-    } catch (e) {
-      console.log("🚀 ~ file: controller.ts ~ line 195 ~ e", e);
-      const token = await connectionSchema[item].exchangeToken(refresh_token);
+    }
+
+    for (let i = 0; i < types.length; i++) {
+      const item: keyof typeof Services = types[i];
+      const { access_token, refresh_token, id } = integrations.find(
+        (integration) => integration.type === Services[item]
+      ) as Integration;
 
       try {
         const event: any = await connectionSchema[item].createEvent(
-          token?.data.access_token,
+          access_token,
           {
             description,
             endTime,
@@ -223,8 +212,8 @@ export const CreateEvents = async (
 
         const newEvent = new Event({
           subject: normalizedSubject,
-          start_time: moment.utc(start.dateTime).valueOf(),
-          end_time: moment.utc(end.dateTime).valueOf(),
+          start_time: moment(start.dateTime).valueOf(),
+          end_time: moment(end.dateTime).valueOf(),
           user_id: userId,
           integration_id: id,
           body: "",
@@ -233,12 +222,49 @@ export const CreateEvents = async (
 
         await newEvent.save();
       } catch (e) {
-        console.log("🚀 ~ file: controller.ts ~ line 220 ~ e", e);
-        return next(ApiError.internal("Internal error"));
+        const token = await connectionSchema[item].exchangeToken(refresh_token);
+
+        try {
+          const event: any = await connectionSchema[item].createEvent(
+            token?.data.access_token,
+            {
+              description,
+              endTime,
+              startTime,
+              subject,
+            }
+          );
+
+          const normalizedEvent = connectionSchema[item].normalize(event.data);
+
+          const {
+            end,
+            id: eventId,
+            start,
+            subject: normalizedSubject,
+          } = normalizedEvent;
+
+          const newEvent = new Event({
+            subject: normalizedSubject,
+            start_time: moment.utc(start.dateTime).valueOf(),
+            end_time: moment.utc(end.dateTime).valueOf(),
+            user_id: userId,
+            integration_id: id,
+            body: "",
+            event_id: eventId,
+          } as any);
+
+          await newEvent.save();
+        } catch (e) {
+          return next(ApiError.internal("Internal error"));
+        }
       }
     }
+    return res.json({ status: "ok" });
+  } catch (e) {
+    console.log("🚀 ~ file: controller.ts ~ line 265 ~ e", e);
+    return next(ApiError.internal("Internal Error"));
   }
-  return res.json({ status: "ok" });
 };
 
 export const DeleteEvent = async (
@@ -253,24 +279,34 @@ export const DeleteEvent = async (
   const event = await Event.findOne({
     where: {
       user_id: id,
-      // id: eventId,
       event_id: eventId,
     },
   });
 
-  const integration = await Integration.findOne({
-    where: {
-      id: event?.integration_id,
-    },
-  });
-  console.log("🚀 ~ file: controller.ts ~ line 267 ~ integration", integration);
+  let integration;
+
+  if (event?.integration_id) {
+    integration = await Integration.findOne({
+      where: {
+        id: event?.integration_id,
+      },
+    });
+  } else {
+    await Event.destroy({
+      where: {
+        user_id: id,
+        event_id: eventId,
+      },
+    });
+
+    return res.json({ status: "ok" });
+  }
 
   try {
-    const deletedEvent = await Event.destroy({
+    await Event.destroy({
       where: {
-        // id: eventId,
-        event_id: eventId,
         user_id: id,
+        event_id: eventId,
       },
     });
   } catch (e) {
@@ -304,7 +340,6 @@ export const DeleteEvent = async (
         eventId: eventId,
       });
     } catch (e: any) {
-      console.log("🚀 ~ file: controller.ts ~ line 306 ~ e", e.message);
       return next(ApiError.badRequest("Internal error"));
     }
   }
